+++ module/item/item-sheet.mjs
@@ -655,11 +655,15 @@ export class HeroSystem6eItemSheet extends FoundryVttItemSheet {
 
     async _onConvertToPower(event) {
         event.preventDefault();
-        await this.item.update({ type: "power" });
+
+        // Also need to use force replace ==items for this to work in v13
+        await this.item.update({ [`type`]: "power", [`==system`]: this.item.system }, { recursive: false });
     }
 
     async _onConvertToEquipment(event) {
         event.preventDefault();
-        await this.item.update({ type: "equipment" });
+
+        // Also need to use force replace ==items for this to work in v13
+        await this.item.update({ [`type`]: "equipment", [`==system`]: this.item.system }, { recursive: false });
     }
 }